adding tests for accconst useconst and status widgets, testing for both to and from xml
