+++ tests/test_contactinfo_gui.py
@@ -15,7 +15,7 @@ def test_contactinfo__from_xml(qtbot):
     widget = ContactInfo.ContactInfo()
     qtbot.addWidget(widget)
 
-    test_record_fname = "C:/Users/mhannon/dev_mdwizard/pymdwizard/tests/data/Onshore_Industrial_Wind_Turbine_Locations_for_the_United_States_through_July2013.xml"
+    test_record_fname = "tests/data/Onshore_Industrial_Wind_Turbine_Locations_for_the_United_States_through_July2013.xml"
     test_record = etree.parse(test_record_fname)
     contact = test_record.xpath("idinfo/ptcontac/cntinfo")[0]
 
@@ -25,4 +25,3 @@ def test_contactinfo__from_xml(qtbot):
     assert widget.findChild(QLineEdit, 'cntvoice').text() == "303-236-5369"
     assert widget.findChild(QLineEdit, 'cntemail').text() == "jediffendorfer@usgs.gov"
     assert widget.findChild(QLineEdit, 'postal').text() == "80225"
-
