+++ tests/test_useconstraints.py
@@ -15,9 +15,20 @@ def test_useconstraints__from_xml(qtbot):
     widget = UseConstraints.UseConstraints()
     qtbot.addWidget(widget)
 
-    test_record_fname = "C:/Users/mhannon/dev_mdwizard/pymdwizard/tests/data/Onshore_Industrial_Wind_Turbine_Locations_for_the_United_States_through_July2013.xml"
+    test_record_fname = "tests/data/Onshore_Industrial_Wind_Turbine_Locations_for_the_United_States_through_July2013.xml"
     test_record = etree.parse(test_record_fname)
-    use_const = test_record.xpath("idinfo")[0]
+    use_const = test_record.xpath("idinfo/useconst")[0]
 
     widget._from_xml(use_const)
-    assert widget.findChild(QPlainTextEdit, "useconst").text() == 'none'
+    assert widget.findChild(QPlainTextEdit, "fgdc_useconst").toPlainText() == 'none'
+
+def test_useconstraints__to_xml(qtbot):
+    widget = UseConstraints.UseConstraints()
+    qtbot.addWidget(widget)
+
+    #test_record_fname = "C:/Users/mhannon/dev_mdwizard/pymdwizard/tests/data/Onshore_Industrial_Wind_Turbine_Locations_for_the_United_States_through_July2013.xml"
+    #test_record = etree.parse(test_record_fname)
+    #use_const = test_record.xpath("idinfo/useconst")[0]
+
+    widget._to_xml#(use_const)
+    assert widget.findChild(QPlainTextEdit, "fgdc_useconst").toPlainText() == "None.  Users are advised to read the data set's metadata thoroughly to understand appropriate use and data limitations."
